+++ examples/tutorial_derive/03_01_flag_count.rs
@@ -3,8 +3,8 @@ use clap::Parser;
 #[derive(Parser)]
 #[clap(author, version, about, long_about = None)]
 struct Cli {
-    #[clap(short, long, parse(from_occurrences))]
-    verbose: usize,
+    #[clap(short, long, action = clap::ArgAction::Count)]
+    verbose: u64,
 }
 
 fn main() {
